+++ CHANGES.rst
@@ -17,6 +17,7 @@ Changelog
 
 **Fixed**
 
+- #1581 Fix Some values are not properly rendered in services listing
 - #1580 Fix Analysts are not displayed once created in worksheets listing
 - #1575 Fix Uncertainties are displayed although result is below Detection Limit
 - #1572 Fix Unable to get the previous status when duplicated in review history